+++ packages/uploadthing/src/_internal/deprecations.ts
@@ -1,0 +1,4 @@
+export const logDeprecationWarning = (message: string) => {
+  // eslint-disable-next-line no-console
+  console.warn(`⚠️ [uploadthing] ${message}`);
+};
